+++ v3/changelog.md
@@ -6,6 +6,7 @@
 
 - Added configuration for headers in response to client for requests handled by
   the pre-route plugin hook.
+- Added OpenAPI schema for filter parameters in JSONAPI
 
 ### Fixed